Move the tests into their own folder
